requirements.txt
# Core ML
tensorflow==2.13.1

# Data handling
numpy>=1.23
pandas>=1.5

# Image processing
pillow>=9.5

# App framework
streamlit>=1.36

# Environment variables
python-dotenv>=1.0

# Database
mysql-connector-python>=8.0

# Optional but recommended for progress bars, downloads, etc.
requests>=2.31
tqdm>=4.66
